+++ src/container/inversify.config.ts
@@ -12,6 +12,7 @@ import { AdvancedModelAnalyzer } from '../analyzers/advancedModelAnalyzer';
 import { UrlPatternAnalyzer } from '../analyzers/urlPatternAnalyzer';
 import { DjangoFormAnalyzer } from '../analyzers/djangoFormAnalyzer';
 import { ViewContextAnalyzer } from '../analyzers/viewContextAnalyzer';
+import { StaticFileAnalyzer } from '../analyzers/staticFileAnalyzer';
 
 // Configuration
 import { ProjectPathConfigurator } from '../projectPathConfigurator';
@@ -27,6 +28,7 @@ import { UrlTagCompletionProvider } from '../providers/urlTagCompletionProvider'
 import { DjangoFormsCompletionProvider } from '../providers/djangoFormsCompletionProvider';
 import { DjangoModelFormCompletionProvider } from '../providers/djangoModelFormCompletionProvider';
 import { TemplateContextCompletionProvider } from '../providers/templateContextCompletionProvider';
+import { StaticPathCompletionProvider } from '../providers/staticPathCompletionProvider';
 
 // Definition providers
 import { DjangoDefinitionProvider } from '../providers/djangoDefinitionProvider';
@@ -60,6 +62,7 @@ export function createContainer(context: vscode.ExtensionContext): Container {
     container.bind<UrlPatternAnalyzer>(TYPES.UrlPatternAnalyzer).to(UrlPatternAnalyzer).inSingletonScope();
     container.bind<DjangoFormAnalyzer>(TYPES.DjangoFormAnalyzer).to(DjangoFormAnalyzer).inSingletonScope();
     container.bind<ViewContextAnalyzer>(TYPES.ViewContextAnalyzer).to(ViewContextAnalyzer).inSingletonScope();
+    container.bind<StaticFileAnalyzer>(TYPES.StaticFileAnalyzer).to(StaticFileAnalyzer).inSingletonScope();
     
     // Configuration - Singleton
     container.bind<ProjectPathConfigurator>(TYPES.ProjectPathConfigurator).to(ProjectPathConfigurator).inSingletonScope();
@@ -76,6 +79,7 @@ export function createContainer(context: vscode.ExtensionContext): Container {
     container.bind<DjangoFormsCompletionProvider>(TYPES.DjangoFormsCompletionProvider).to(DjangoFormsCompletionProvider);
     container.bind<DjangoModelFormCompletionProvider>(TYPES.DjangoModelFormCompletionProvider).to(DjangoModelFormCompletionProvider);
     container.bind<TemplateContextCompletionProvider>(TYPES.TemplateContextCompletionProvider).to(TemplateContextCompletionProvider);
+    container.bind<StaticPathCompletionProvider>(TYPES.StaticPathCompletionProvider).to(StaticPathCompletionProvider);
     
     // Definition providers - Transient
     container.bind<DjangoDefinitionProvider>(TYPES.DjangoDefinitionProvider).to(DjangoDefinitionProvider);